requirements.txt
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.24
pandas>=2.0
shapely>=2.0
pyproj>=3.6
mercantile>=1.2
Pillow>=10.0
PyYAML>=6.0
folium>=0.15
matplotlib>=3.7
